add getting actual rates of currency from i-net, parsing response from server and saving rates in app

+++ src/main/java/org/dng/springApi_2022_10_29/DTO/IMessage.java
@@ -3,5 +3,5 @@ package org.dng.springApi_2022_10_29.DTO;
 public interface IMessage {
     Messages.Currency inCurrency();
     Messages.Currency outCurrency();
-    float value();
+    double value();
 }
