+++ src/main/java/org/dng/springApi_2022_10_29/DTO/Messages.java
@@ -3,16 +3,16 @@ package org.dng.springApi_2022_10_29.DTO;
 
 public class Messages {
 
-    public record InMessage(Currency inCurrency, Currency outCurrency, float value) implements IMessage {
-        public InMessage(Currency inCurrency, Currency outCurrency, float value) {
+    public record InMessage(Currency inCurrency, Currency outCurrency, double value) implements IMessage {
+        public InMessage(Currency inCurrency, Currency outCurrency, double value) {
             this.inCurrency = inCurrency;
             this.outCurrency = outCurrency;
             this.value = value;
         }
     }
 
-    public record OutMessage(Currency inCurrency, Currency outCurrency, float value) implements IMessage {
-        public OutMessage(Currency inCurrency, Currency outCurrency, float value) {
+    public record OutMessage(Currency inCurrency, Currency outCurrency, double value) implements IMessage {
+        public OutMessage(Currency inCurrency, Currency outCurrency, double value) {
             this.inCurrency = inCurrency;
             this.outCurrency = outCurrency;
             this.value = value;
